添加版本更新提示
1、删除部分无用图片
2、添加版本更新提示

+++ myoffer/MyOffer/Class/ApplyList/controller/ApplyViewController.m
@@ -386,7 +386,7 @@ typedef NS_ENUM(NSInteger,ApplyTableStatus){
    
         
         //如果学校被选择状态，在先中cell时，要删除学校的选中状态
-        NSString *sectionStr = [NSString stringWithFormat:@"%ld",indexPath.section];
+        NSString *sectionStr = [NSString stringWithFormat:@"%ld",(long)indexPath.section];
         if ([self.cancelSetions containsObject: sectionStr]) {
             
             [self.cancelSetions removeObject:sectionStr];
